WIP: Dockerfile and makefile for nginx container
The docker file should be complete, I'm working on the NGINX
configuration.  So far I've got the UI up and running, but it's not
connecting to the API yet.  I've figure out that I can use
environment variables to configure the API URL, but I'm not sure how
to get the location blocks correct in the NGINX configuration.

I've copied the proxy configuration from the `vue.config.js` file, but
something's still not quite right.